RICHEDTI20 uses character rather than byte as the unit when searching an selecting. When the code page of your system contains multi-byte characters, and you used them as the symbolic name or comment, the return value of SendDlgItemMessage() can't be referenced to the index in char* strings, this problem can make all the highlighting in the disassembly list screwed up. I don't know how many people would like to symbolic name or comment the code with their own language, and most people on the Earth may never encounter this problem because the code page of their system language is all single-byte. However, it's quite annoying when you are not good at English and your language contains many MBCs. As the whole FCEUX project is in ANSI, it's impractical to change all the strings in the project to Unicode, but since it accepts multi-byte characters in most places without big problems, this should be fixed in some way. I have tested this fix in Windows VM of several languages, it currently works fine, although converting the whole debugging text to wchar_t every time is not very efficient.

+++ src/drivers/win/memview.cpp
@@ -1542,7 +1542,7 @@ LRESULT CALLBACK MemViewCallB(HWND hwnd, UINT message, WPARAM wParam, LPARAM lPa
 			if (wParam >= '0' && wParam <= '9')
 			{
 				char buf[3];
-				sprintf(buf, "%c", wParam);
+				sprintf(buf, "%c", (int)wParam);
 				int key_num;
 				sscanf(buf, "%d", &key_num);
 				key_num = (key_num + 9) % 10;
